+++ pkg/resources/resource_connection_kafka_test.go
@@ -13,11 +13,17 @@ import (
 )
 
 var inKafka = map[string]interface{}{
-	"name":                      "conn",
-	"schema_name":               "schema",
-	"database_name":             "database",
-	"service_name":              "service",
-	"kafka_broker":              []interface{}{map[string]interface{}{"broker": "b-1.hostname-1:9096", "target_group_port": 9001, "availability_zone": "use1-az1", "privatelink_conn": "privatelink_conn"}},
+	"name":          "conn",
+	"schema_name":   "schema",
+	"database_name": "database",
+	"service_name":  "service",
+	"kafka_broker": []interface{}{map[string]interface{}{
+		"broker":                 "b-1.hostname-1:9096",
+		"target_group_port":      9001,
+		"availability_zone":      "use1-az1",
+		"privatelink_connection": []interface{}{map[string]interface{}{"name": "cert"}},
+		"ssh_tunnel":             []interface{}{map[string]interface{}{"name": "ssh"}},
+	}},
 	"security_protocol":         "SASL_PLAINTEXT",
 	"progress_topic":            "topic",
 	"ssl_certificate_authority": []interface{}{map[string]interface{}{"text": "key"}},
@@ -39,7 +45,20 @@ func TestResourceConnectionKafkaCreate(t *testing.T) {
 	testhelpers.WithMockDb(t, func(db *sqlx.DB, mock sqlmock.Sqlmock) {
 		// Create
 		mock.ExpectExec(
-			`CREATE CONNECTION "database"."schema"."conn" TO KAFKA \(BROKERS \('b-1.hostname-1:9096' USING SSH TUNNEL "materialize"."public"."tunnel"\), SECURITY PROTOCOL = 'SASL_PLAINTEXT', PROGRESS TOPIC 'topic', SSL CERTIFICATE AUTHORITY = 'key', SSL CERTIFICATE = SECRET "materialize"."public"."cert", SSL KEY = SECRET "materialize"."public"."key", SASL MECHANISMS = 'PLAIN', SASL USERNAME = 'username', SASL PASSWORD = SECRET "materialize"."public"."password"\);`,
+			`CREATE CONNECTION "database"."schema"."conn"
+			TO KAFKA \(BROKERS
+				\('b-1.hostname-1:9096'
+				USING SSH TUNNEL "materialize"."public"."ssh"
+				USING AWS PRIVATELINK "materialize"."public"."cert" 
+				\(PORT 9001, AVAILABILITY ZONE 'use1-az1'\)\),
+			SSH TUNNEL "materialize"."public"."tunnel",
+			SECURITY PROTOCOL = 'SASL_PLAINTEXT', PROGRESS TOPIC 'topic',
+			SSL CERTIFICATE AUTHORITY = 'key',
+			SSL CERTIFICATE = SECRET "materialize"."public"."cert",
+			SSL KEY = SECRET "materialize"."public"."key",
+			SASL MECHANISMS = 'PLAIN',
+			SASL USERNAME = 'username',
+			SASL PASSWORD = SECRET "materialize"."public"."password"\);`,
 		).WillReturnResult(sqlmock.NewResult(1, 1))
 
 		// Comment